database php code
A bug with updating entries needs to be sorted out as well when making
an entry with a large amount of text. The displayartists page needs to
only show a preview of the artist details. The details page still needs
to be coded so that it only shows one artist based on which link you
clicked.